+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Added
 
-- Add `org.jetbrains.intellij.deps.kotlinx` as a `kotlinx.coroutines` shadowed dependency group eligible for exclusion. 
+- Add `org.jetbrains.intellij.deps.kotlinx` as a `kotlinx.coroutines` shadowed dependency group eligible for exclusion.
 - Expand `TestFrameworkType` with additional frameworks including Debugger, ExternalSystem, PolySymbols, Qodana, UAST, VCS, and WebSymbols.
+- Introduce `org.jetbrains.intellij.platform.intellijPlatformIdesCacheEnabled` Gradle property to control IntelliJ Platform IDEs caching globally via `intellijPlatform.caching.ides.enabled` convention.
 
 ## [2.10.4] - 2025-11-02